SimuUtils

Documentation for SimuUtils.

SimuUtils.UOM_CONVERSIONS

SimuUtils.Casing

SimuUtils.area_circle

SimuUtils.calc_casing_id

SimuUtils.calc_casing_weight

SimuUtils.calculate_r_squared

SimuUtils.convert

SimuUtils.convert

SimuUtils.convert

SimuUtils.linear_regression

SimuUtils.non_linear_regression

SimuUtils.plot_regression

SimuUtils.plot_well_schem

SimuUtils.round_to

SimuUtils.UOM_CONVERSIONS — Constant

UOM_CONVERSIONS

A dictionary of conversion factors between different units of measure.

SimuUtils.Casing — Method

Casing(
	od::T, id::T, units::DiameterUnit=IN::DiameterUnit;
	hanger::Union{Nothing, Tuple{T, LengthUnit}}=nothing,
	depth::Union{Nothing, Tuple{T, LengthUnit}}=nothing) where T<:Union{Float64, Float32}

Constructor given od and id inputs. depth and other Casing properties are optional.

Arguments

od::T: The outer diameter of the casing [in or mm].

id::T: The inner diameter of the casing [in or mm].

units::DiameterUnit: The unit of measurement for the diameter inputs.

depth::Union{Nothing, Tuple{T, LengthUnit}}: The depth of the casing [ft or m].

Result

Casing{T}: The constructed Casing object.

SimuUtils.area_circle — Method

area_circle(od::T, id::T=0.0; xs::T=0.0) where T <: Real

Calculate the area of a ring given the outer diameter and inner diameter. If inner diameter is not provided, it is assumed to be zero. It also accepts a cross-sectional % excess area.

Arguments

od::T: The outer diameter of the ring.

id::T=0.0: The inner diameter of the ring.

xs::T=0.0: The cross-sectional % excess area.

Returns

T: The area of the ring.

SimuUtils.calc_casing_id — Method

calc_casing_id(
    od::T, weight_ppf::T, unit::DiameterUnit;
    material::T=STEEL_DENSITY) where T<:Union{Float64, Float32}

Function to calculate the inner diameter of the casing given the outer diameter and weight in ppf.

Arguments

od::T: Outer diameter of the casing [in or mm].

weight_ppf::T: The weight of the casing in lb/ft.

unit::DiameterUnit: The unit of measurement for the diameters.

material::T=STEEL_DENSITY: The density of the material in lb/ft^3. Defaults to STEEL_DENSITY.

Returns

id::T: The inner diameter of the casing [in or mm].

SimuUtils.calc_casing_weight — Method

calculate_casing_weight(
	od::T, id::T, unit::DiameterUnit;
	material::T=STEEL_DENSITY) where T<:Union{Float64, Float32}

Function to calculate the weight of the casing in ppf (lb/ft)

Arguments

od::T: Outer diameter of the casing [in or mm].

id::T: Inner diameter of the casing [in or mm].

material::T=STEEL_DENSITY: The density of the material in lb/ft^3. Defaults to STEEL_DENSITY.

Returns

weight_ppf::T: The weight of the casing in lb/ft.

SimuUtils.calculate_r_squared — Method

calculate_r2(Y::Vector{Float64}, Y_fit::Vector{Float64})

This function calculates the R^2 value for the given Y and Y_fit data points.

Arguments

Y::Vector{Float64}: The original Y-axis data points.

Y_fit::Vector{Float64}: The fitted Y-axis data points.

Returns

Float64: The R^2 value.

SimuUtils.convert — Method

convert(value::T, from_unit::U, to_unit::U) where {T<:Number, U<:Unit}

Convert a value from one unit of measure to another.

Arguments

value::T: The value to convert.

from_unit::U: The unit of measure to convert from.

to_unit::U: The unit of measure to convert to.

Returns

T: The converted value.

SimuUtils.convert — Method

convert(uom::Uom{T, U}, to_unit::U) where {T, U}

Convert a Uom struct from one unit of measure to another.

Arguments

uom::Uom{T, U}: The Uom struct to convert.

to_unit::U: The unit of measure to convert to.

Returns

Uom{T, U}: The converted Uom struct.

SimuUtils.convert — Method

convert(value::T, from_unit::U, to_unit::U) where {T<:Number, U}

Convert a value from one unit of measure to another.

Arguments

value::T: The value to convert.

from_unit::U: The unit of measure to convert from.

to_unit::U: The unit of measure to convert to.

Returns

T: The converted value.

SimuUtils.linear_regression — Method

linear_regression(X::Vector{Float64}, Y::Vector{Float64})

This function calculates the linear regression coefficients (slope and intercept)
for the given X and Y data points.

Arguments

X::Vector{Float64}: The X-axis data points.

Y::Vector{Float64}: The Y-axis data points.

Returns

Tuple{Float64, Float64}: The linear regression coefficients (slope, intercept).

SimuUtils.non_linear_regression — Method

non_linear_regression(X::Vector{Float64}, Y::Vector{Float64})

This function calculates the non-linear regression coefficients (a, b, c)
for the given X and Y data points using a quadratic model.

Arguments

X::Vector{Float64}: The X-axis data points.

Y::Vector{Float64}: The Y-axis data points.

Returns

Vector{Float64}: The non-linear regression coefficients (a, b, c).

SimuUtils.plot_regression — Method

linear_regression(X::Vector{Float64}, Y::Vector{Float64})

plot_regression(
	X_exp::Vector{Float64}, Y_exp::Vector{Float64},
	X_fit::Vector{Float64}, Y_fit::Vector{Float64},
	regression_type::Symbol;
	title::String="Regression Plot", xlabel::String="x-axis", ylabel::String="y-axis",
	annotations::Dict{String, Float64}=Dict(), figres=(450, 300)
)

This function plots the experimental data points along with
the fitted model data using linear or non-linear regression.

Arguments

X_exp: The experimental X-axis data points.

Y_exp: The experimental Y-axis data points.

X_fit: The fitted X-axis data points.

Y_fit: The fitted Y-axis data points.

regression_type: The type of regression to use (:linear or :non_linear).

title: The title of the plot.

xlabel: The label for the x-axis.

ylabel: The label for the y-axis.

annotations: The annotations to display on the plot.

figres: The resolution of the figure.

scale: Define the scale of Axis (e.g. idenntitiy, log10, log2)

Returns

Nothing

SimuUtils.plot_well_schem — Method

plot_well_schem(;
	outer_geometry::Vector{Casing},
	inner_geometry::Vector{Casing};
	airgap=nothing, water_depth=nothing,
	units::LengthUnit=M::LengthUnit)

Plots a 2D wellbore schematic with outer and inner casings.

SimuUtils.round_to — Method

round_to(value::T, decimals::Int) where T <: Real

Round a value to a given number of decimals.

Arguments

value::T: The value to round.

decimals::Int: The number of decimals to round to.

Returns

T: The rounded value.